backend/requirements.txt
fastapi[all] >=0.89.1, <0.90.0
pytest >=7.2.1, <7.3.0
psycopg2 >=2.9.5, <2.10.0
sqlalchemy >=2.0.4, <2.1.0
python-dotenv >=1.0.0, <1.1.0